move settings update tests back
