+++ crates/settings/src/settings_json.rs
@@ -352,207 +352,4 @@ pub fn parse_json_with_comments<T: DeserializeOwned>(content: &str) -> Result<T>
 }
 
 #[cfg(test)]
-mod tests {
-    use gpui::App;
-    use unindent::Unindent;
-
-    use crate::{
-        Settings, SettingsStore,
-        settings_store::tests::{
-            LanguageSettingEntry, LanguageSettings, MultiKeySettings, UserSettings,
-        },
-    };
-
-    fn check_settings_update<T: Settings>(
-        store: &mut SettingsStore,
-        old_json: String,
-        update: fn(&mut T::FileContent),
-        expected_new_json: String,
-        cx: &mut App,
-    ) {
-        store.set_user_settings(&old_json, cx).ok();
-        let edits = store.edits_for_update::<T>(&old_json, update);
-        let mut new_json = old_json;
-        for (range, replacement) in edits.into_iter() {
-            new_json.replace_range(range, &replacement);
-        }
-        pretty_assertions::assert_eq!(new_json, expected_new_json);
-    }
-
-    fn check_keymap_update() {}
-
-    #[gpui::test]
-    fn test_setting_store_update(cx: &mut App) {
-        let mut store = SettingsStore::new(cx);
-        store.register_setting::<MultiKeySettings>(cx);
-        store.register_setting::<UserSettings>(cx);
-        store.register_setting::<LanguageSettings>(cx);
-
-        // entries added and updated
-        check_settings_update::<LanguageSettings>(
-            &mut store,
-            r#"{
-                     "languages": {
-                         "JSON": {
-                             "language_setting_1": true
-                         }
-                     }
-                 }"#
-            .unindent(),
-            |settings| {
-                settings
-                    .languages
-                    .get_mut("JSON")
-                    .unwrap()
-                    .language_setting_1 = Some(false);
-                settings.languages.insert(
-                    "Rust".into(),
-                    LanguageSettingEntry {
-                        language_setting_2: Some(true),
-                        ..Default::default()
-                    },
-                );
-            },
-            r#"{
-                     "languages": {
-                         "Rust": {
-                             "language_setting_2": true
-                         },
-                         "JSON": {
-                             "language_setting_1": false
-                         }
-                     }
-                 }"#
-            .unindent(),
-            cx,
-        );
-
-        // entries removed
-        check_settings_update::<LanguageSettings>(
-            &mut store,
-            r#"{
-                     "languages": {
-                         "Rust": {
-                             "language_setting_2": true
-                         },
-                         "JSON": {
-                             "language_setting_1": false
-                         }
-                     }
-                 }"#
-            .unindent(),
-            |settings| {
-                settings.languages.remove("JSON").unwrap();
-            },
-            r#"{
-                     "languages": {
-                         "Rust": {
-                             "language_setting_2": true
-                         }
-                     }
-                 }"#
-            .unindent(),
-            cx,
-        );
-
-        check_settings_update::<LanguageSettings>(
-            &mut store,
-            r#"{
-                     "languages": {
-                         "Rust": {
-                             "language_setting_2": true
-                         },
-                         "JSON": {
-                             "language_setting_1": false
-                         }
-                     }
-                 }"#
-            .unindent(),
-            |settings| {
-                settings.languages.remove("Rust").unwrap();
-            },
-            r#"{
-                     "languages": {
-                         "JSON": {
-                             "language_setting_1": false
-                         }
-                     }
-                 }"#
-            .unindent(),
-            cx,
-        );
-
-        // weird formatting
-        check_settings_update::<UserSettings>(
-            &mut store,
-            r#"{
-                     "user":   { "age": 36, "name": "Max", "staff": true }
-                 }"#
-            .unindent(),
-            |settings| settings.age = Some(37),
-            r#"{
-                     "user":   { "age": 37, "name": "Max", "staff": true }
-                 }"#
-            .unindent(),
-            cx,
-        );
-
-        // single-line formatting, other keys
-        check_settings_update::<MultiKeySettings>(
-            &mut store,
-            r#"{ "one": 1, "two": 2 }"#.unindent(),
-            |settings| settings.key1 = Some("x".into()),
-            r#"{ "key1": "x", "one": 1, "two": 2 }"#.unindent(),
-            cx,
-        );
-
-        // empty object
-        check_settings_update::<UserSettings>(
-            &mut store,
-            r#"{
-                     "user": {}
-                 }"#
-            .unindent(),
-            |settings| settings.age = Some(37),
-            r#"{
-                     "user": {
-                         "age": 37
-                     }
-                 }"#
-            .unindent(),
-            cx,
-        );
-
-        // no content
-        check_settings_update::<UserSettings>(
-            &mut store,
-            r#""#.unindent(),
-            |settings| settings.age = Some(37),
-            r#"{
-                     "user": {
-                         "age": 37
-                     }
-                 }
-                 "#
-            .unindent(),
-            cx,
-        );
-
-        check_settings_update::<UserSettings>(
-            &mut store,
-            r#"{
-                 }
-                 "#
-            .unindent(),
-            |settings| settings.age = Some(37),
-            r#"{
-                     "user": {
-                         "age": 37
-                     }
-                 }
-                 "#
-            .unindent(),
-            cx,
-        );
-    }
-}
+mod tests {}
